+++ Demonstrativo/Models/ItemVendaViewModel.cs
@@ -10,7 +10,7 @@ namespace Demonstrativo.Models
         public int Id { get; set; }
         public int VendaId { get; set; }
         public int ProdutoId { get; set; }
-        public double Quantidade { get; set; }
+        public decimal Quantidade { get; set; }
         public decimal Preco { get; set; }
         public ProdutoViewModel Produto { get; set; }
     }
